# Optimized __method__ and __callee__