--- a/core/src/main/java/org/jruby/internal/runtime/methods/InterpretedIRBodyMethod.java
+++ b/core/src/main/java/org/jruby/internal/runtime/methods/InterpretedIRBodyMethod.java
@@ -71,7 +71,7 @@ public class InterpretedIRBodyMethod extends InterpretedIRMethod {
     private IRubyObject interpretWithBacktrace(InterpreterContext ic, ThreadContext context, IRubyObject self, String name, Block block) {
         try {
             ThreadContext.pushBacktrace(context, name, ic.getFileName(), ic.getLine());
-            return ic.getEngine().interpret(context, null, self, ic, getImplementationClass().getMethodLocation(), name, block);
+            return ic.getEngine().interpret(context, null, self, ic, getImplementationClass().getMethodLocation(), null, block);
         } finally {
             ThreadContext.popBacktrace(context);
         }