--- a/core/src/main/java/org/jruby/ir/instructions/FrameNameCallInstr.java
+++ b/core/src/main/java/org/jruby/ir/instructions/FrameNameCallInstr.java
@@ -56,6 +56,8 @@ public class FrameNameCallInstr extends NoOperandResultBaseInstr implements Fixe
             return frameNameSite.call(context, self, self);
         }
 
+        if (compositeName == null) return context.nil;
+
         return callee ?
                 RubySymbol.newCalleeSymbolFromCompound(context.runtime, compositeName):
                 RubySymbol.newMethodSymbolFromCompound(context.runtime, compositeName);