--- a/core/src/main/java/org/jruby/ir/interpreter/ExitableInterpreterEngine.java
+++ b/core/src/main/java/org/jruby/ir/interpreter/ExitableInterpreterEngine.java
@@ -108,7 +108,7 @@ public class ExitableInterpreterEngine extends InterpreterEngine {
                         break;
                     case CALL_OP:
                         if (profile) Profiler.updateCallSite(instr, interpreterContext.getScope(), scopeVersion);
-                        processCall(context, instr, operation, currDynScope, currScope, temp, self, name);
+                        processCall(context, instr, operation, currDynScope, currScope, temp, self, name, block);
                         break;
                     case RET_OP:
                         processReturnOp(context, block, instr, operation, currDynScope, temp, self, currScope);