… bench/bench_frame_name_methods.rb
… core/src/main/java/org/jruby/ir/targets/indy/IndyInvocationCompiler.java
… core/src/main/java/org/jruby/ir/targets/simple/NormalInvocationCompiler.java