@headius 2024-03-28
Additional optimizations:

* Use a record to hold the symbol pair.
* Cache the last-seen simple method symbol in the site and use it when appropriate.
* Inject the symbol table into the indy handle chain to avoid repeatedly acquiring it.

With all changes, final bench numbers look like this:

```
[] jruby $ jruby -Xcompile.invokedynamic bench/bench_frame_name_methods.rb                         
Warming up --------------------------------------
     __method__ same     5.484M i/100ms
__method__ different     5.391M i/100ms
     __callee__ same     4.506M i/100ms
__callee__ different     4.501M i/100ms
Calculating -------------------------------------
     __method__ same     57.464M (± 0.7%) i/s -    290.649M in   5.058218s
__method__ different     57.374M (± 0.9%) i/s -    291.106M in   5.074304s
     __callee__ same     45.069M (± 1.0%) i/s -    225.302M in   4.999640s
__callee__ different     45.042M (± 1.4%) i/s -    225.070M in   4.997901s                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        [] jruby $ jruby bench/bench_frame_name_methods.rb
Warming up --------------------------------------
     __method__ same     1.620M i/100ms
__method__ different   943.069k i/100ms
     __callee__ same     1.604M i/100ms
__callee__ different   960.143k i/100ms
Calculating -------------------------------------
     __method__ same     16.377M (± 1.6%) i/s -     82.611M in   5.045839s
__method__ different      9.156M (± 1.3%) i/s -     46.210M in   5.047853s
     __callee__ same     15.847M (± 1.8%) i/s -     80.219M in   5.063688s
__callee__ different      9.183M (± 2.0%) i/s -     46.087M in   5.021116s
```

Note that the slow compound name path appears to have some volatility in the JVM JIT, and occasionally now drops about 25% of its performance:

```
[] jruby $ jruby bench/bench_frame_name_methods.rb 
Warming up --------------------------------------
     __method__ same     1.648M i/100ms
__method__ different   764.744k i/100ms
     __callee__ same     1.622M i/100ms
__callee__ different   754.497k i/100ms
Calculating -------------------------------------
     __method__ same     16.376M (± 2.2%) i/s -     82.403M in   5.034454s
__method__ different      7.017M (± 4.3%) i/s -     35.178M in   5.022898s
     __callee__ same     15.908M (± 6.5%) i/s -     79.478M in   5.034325s
__callee__ different      7.137M (± 0.7%) i/s -     36.216M in   5.074376s       
```

This may be due to the additional optimizations confounding HotSpot *sometimes*, or it may have always been there but not during the benchmark runs I performed.